android: move files around
