+++ android/app/src/main/java/io/github/parkwithease/parkeasy/data/Credentials.kt
@@ -1,4 +1,4 @@
-package io.github.parkwithease.parkeasy
+package io.github.parkwithease.parkeasy.data
 
 import kotlinx.serialization.Serializable
