+++ android/app/src/main/java/io/github/parkwithease/parkeasy/ui/login/LoginViewModel.kt
@@ -1,0 +1,5 @@
+package io.github.parkwithease.parkeasy.ui.login
+
+class LoginViewModel {
+
+}
